@@ -1,6 +1,6 @@
 {
   "name": "clrc",
-  "version": "2.2.0",
+  "version": "3.0.0",
   "description": "LRC format parser.",
   "main": "dist/index.js",
   "scripts": {
@@ -30,16 +30,15 @@
   },
   "homepage": "https://github.com/mebtte/clrc#readme",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "eslint": "^7.22.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "^2.22.1",
-    "husky": "^5.1.3",
-    "lint-staged": "^10.5.4",
-    "ora": "^5.3.0",
-    "prettier": "^2.2.1",
-    "typescript": "^4.2.3"
+    "@typescript-eslint/eslint-plugin": "^5.36.1",
+    "@typescript-eslint/parser": "^5.36.1",
+    "eslint": "^8.23.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "husky": "^8.0.1",
+    "lint-staged": "^13.0.3",
+    "prettier": "^2.7.1",
+    "typescript": "^4.8.2"
   }
 }
