@@ -1,6 +1,4 @@
-import ReactDOM from 'react-dom';
-import React from 'react';
-
+import ReactDOM from 'react-dom/client';
 import App from './app';
 
-ReactDOM.render(<App />, document.querySelector('#app'));
+ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
