@@ -1,29 +1,28 @@
-export interface LrcLine {
-  lineNumber: number;
-  raw: string;
+// eslint-disable-next-line no-shadow
+export enum LineType {
+  INVALID = 'invalid',
+  LYRIC = 'lyric',
+  METADATA = 'metadata',
 }
 
-export interface MetadataLine<MetadataKey extends string> extends LrcLine {
+export interface Line {
+  lineNumber: number;
+  raw: string;
+  type: LineType;
+}
+
+export interface MetadataLine<MetadataKey extends string> extends Line {
+  type: LineType.METADATA;
   key: MetadataKey;
   value: string;
 }
 
-export interface LyricLine extends LrcLine {
+export interface LyricLine extends Line {
+  type: LineType.LYRIC;
   startMillisecond: number;
   content: string;
 }
 
-export type ParseOptions = {
-  /** whether to sort lyrics by start time */
-  sortByStartTime?: boolean;
-  /** whether to remove start spaces */
-  trimStart?: boolean;
-  /** whether to remove end spaces */
-  trimEnd?: boolean;
-};
-
-export const DEFAULT_OPTIONS: ParseOptions = {
-  sortByStartTime: false,
-  trimStart: true,
-  trimEnd: false,
-};
+export interface InvalidLine extends Line {
+  type: LineType.INVALID;
+}
